\name{rym-package}
\alias{rym-package}
\alias{rym}
\docType{package}
\title{
\packageTitle{rym}
}
\description{
\packageDescription{rym}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rym}
\packageIndices{rym}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{rym}

Maintainer: \packageMaintainer{rym}
}
\references{
~~ Literature or other references for background information ~~
}

\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
